requirements.txt
discord.py==1.3.2
requests>=2.20.0
lxml==4.5.0
pynacl==1.3.0
firebase-admin==4.2.0
